+++ src/App.vue
@@ -44,7 +44,7 @@
 </template>
 
 <script lang="ts">
-import { Component, Vue, Watch } from 'vue-property-decorator'
+import { Component, Vue } from 'vue-property-decorator'
 
 import aggregatorService from './services/aggregatorService'